[Bronze V] Title: 이상한 기호, Time: 168 ms, Memory: 17740 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -6,8 +6,8 @@ public class Main {
         
         Scanner sc = new Scanner(System.in);
         
-        int A = sc.nextInt();
-        int B = sc.nextInt();
+        Long A = sc.nextLong();
+        Long B = sc.nextLong();
         
         System.out.print((A+B) * (A-B));
     }
